Give map of instances a more dignified name
Move the map of instances in the task manager into its own class. Make
the init of the graph using that data a method of that class. Move these
classes that help encapsulate the capacity and ability of all the instance the task
manager will be assigning work to into another file. It would be great
to add some tests of these classes themselves now that they are discreet
units.
