+++ awx/main/scheduler/task_manager.py
@@ -37,76 +37,10 @@ from awx.main.utils.pglock import advisory_lock
 from awx.main.utils import get_type_for_model, task_manager_bulk_reschedule, schedule_task_manager
 from awx.main.utils.common import create_partition
 from awx.main.signals import disable_activity_stream
-from awx.main.scheduler.dependency_graph import DependencyGraph
 from awx.main.utils import decrypt_field
-
+from awx.main.scheduler.capacity_utils import TaskManagerInstances
 
 logger = logging.getLogger('awx.main.scheduler')
-
-
-class PrioritizableControlNode:
-    def __init__(self, remaining_capacity, hostname):
-        self.remaining_capacity = remaining_capacity
-        self.hostname = hostname
-
-    def __gt__(self, other):
-        return self.remaining_capacity > other.remaining_capacity
-
-    def __lt__(self, other):
-        return self.remaining_capacity < other.remaining_capacity
-
-    def __eq__(self, other):
-        return self.hostname == other.hostname
-
-    def __hash__(self):
-        return hash(self.hostname)
-
-    def __str__(self):
-        return f"{self.hostname}: {self.remaining_capacity} remaining capacity"
-
-    def __repr__(self):
-        return self.__str__()
-
-
-class PrioritizedControlNodes:
-    def __init__(self):
-        self.control_nodes = set()
-
-    def add(self, node):
-        self.control_nodes.add(node)
-
-    def update(self, node):
-        """The __hash__ function for PrioritizableControlNodes just compares hostnames.
-
-        We can add/remove the node to update object in PrioritizedControlNodes to have right capacity.
-        """
-        self.control_nodes.remove(node)
-        self.control_nodes.add(node)
-
-    # TODO: Maybe rename
-    def assign_task_to_node(self, task, instances_partial={}, graph=None):
-        """Find most available control instance and deduct task impact if we find it.
-
-        If no node is available, return False
-        """
-        control_heap = list(self.control_nodes)
-        heapq.heapify(control_heap)
-        best_instance = heapq.nlargest(1, control_heap).pop()
-        if best_instance.remaining_capacity < task.task_impact:
-            return False
-
-        if instances_partial and graph:
-            for ig in instances_partial[best_instance.hostname].instance_groups:
-                ig_data = graph[ig]
-                # This should really not happen since every time we assign a task to a node, we deduct from its instance group
-                # capacity as well. If we never see this, maybe we can drop this code
-                if not ig_data['control_capacity'] - ig_data['consumed_control_capacity'] >= task.task_impact:
-                    logger.warn(f"Somehow we had capacity on a control node {best_instance} but not on its instance_group {instance_group}")
-                    return False
-        logger.debug(f"chose {best_instance} to run {task.log_format}")
-        best_instance.remaining_capacity -= task.task_impact
-        self.update(best_instance)
-        return best_instance
 
 
 class TaskManager:
@@ -134,38 +68,9 @@ class TaskManager:
         """
         Init AFTER we know this instance of the task manager will run because the lock is acquired.
         """
-        instances = Instance.objects.filter(hostname__isnull=False, enabled=True).exclude(node_type='hop').prefetch_related('rampart_groups')
-        self.instances_partial = dict()
-        self.control_nodes = PrioritizedControlNodes()
-        for instance in instances:
-            self.instances_partial[instance.hostname] = SimpleNamespace(
-                obj=instance,
-                node_type=instance.node_type,
-                remaining_capacity=instance.remaining_capacity,
-                capacity=instance.capacity,
-                jobs_running=instance.jobs_running,
-                hostname=instance.hostname,
-                instance_groups=[ig.name for ig in instance.rampart_groups.all()],
-            )
-            if instance.node_type in ('control', 'hybrid'):
-                self.control_nodes.add(PrioritizableControlNode(remaining_capacity=instance.remaining_capacity, hostname=instance.hostname))
-
-        for rampart_group in InstanceGroup.objects.prefetch_related('instances'):
-            self.graph[rampart_group.name] = dict(
-                graph=DependencyGraph(),
-                execution_capacity=0,
-                control_capacity=0,
-                consumed_capacity=0,
-                consumed_control_capacity=0,
-                consumed_execution_capacity=0,
-                instances=[],
-            )
-            for instance in rampart_group.instances.filter(enabled=True).order_by('hostname'):
-                if instance.hostname in self.instances_partial:
-                    self.graph[rampart_group.name]['instances'].append(self.instances_partial[instance.hostname])
-                for capacity_type in ('control', 'execution'):
-                    if instance.node_type in (capacity_type, 'hybrid'):
-                        self.graph[rampart_group.name][f'{capacity_type}_capacity'] += instance.capacity
+        self.task_manager_instances = TaskManagerInstances()
+        self.control_nodes = self.task_manager_instances.control_nodes
+        self.graph = self.task_manager_instances.init_graph(self.graph)
 
     def get_and_consume_capacity_on_control_node_with_sufficient_capacity(self, task):
         """Find a control node with enough capacity to control a job.
@@ -175,14 +80,14 @@ class TaskManager:
 
         Return None if no appropriate instance found.
         """
-        control_node = self.control_nodes.assign_task_to_node(task, self.instances_partial, self.graph)
+        control_node = self.control_nodes.assign_task_to_node(task, self.task_manager_instances, self.graph)
         if not control_node:
             logger.debug(f"No control node eligible to run {task.log_format}")
             return None
 
-        for ig in self.instances_partial[control_node.hostname].instance_groups:
+        for ig in self.task_manager_instances[control_node.hostname].instance_groups:
             # Consume capacity on instance groups the control node is a member of
-            self.consume_capacity(task, ig, instance=self.instances_partial[control_node.hostname])
+            self.consume_capacity(task, ig, instance=self.task_manager_instances[control_node.hostname])
         return control_node.hostname
 
     def job_blocked_by(self, task):
@@ -581,7 +486,7 @@ class TaskManager:
                     controller_node = self.get_and_consume_capacity_on_control_node_with_sufficient_capacity(task)
                     # No controller node was available, go on to next task
                     if not controller_node:
-                        logger.warning("No control plane nodes available to run containerized job {}, returning to pending".format(task.log_format))
+                        logger.warning("No control plane nodes available to run containerized job {}, will remain pending".format(task.log_format))
                         found_acceptable_queue = False
                         continue
                     # If there is enough capacity, assign the node to be the execution node and break out of the sub loop
@@ -595,7 +500,7 @@ class TaskManager:
 
                     # No controller node was available, go on to next task
                     if not controller_node:
-                        logger.warning("No control plane nodes available to run containerized job {}, returning to pending".format(task.log_format))
+                        logger.warning("No control plane nodes available to run containerized job {}, will remain pending".format(task.log_format))
                         found_acceptable_queue = False
                         continue
 
@@ -628,7 +533,7 @@ class TaskManager:
                         )
 
                     if execution_instance:
-                        execution_instance = self.instances_partial[execution_instance.hostname].obj
+                        execution_instance = self.task_manager_instances[execution_instance.hostname].obj
                     self.graph[rampart_group.name]['graph'].add_job(task)
                     self.start_task(task, rampart_group, task.get_jobs_fail_chain(), execution_instance)
                     found_acceptable_queue = True
